Backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
shap-e
huggingface_hub
python-dotenv
psycopg2-binary
stripe
authlib
redis
